Possible correction sans hack.
I believe the intent of the validation in the test is to say "this array
controller is valid when all its children are valid".  I believe this change
achieves those semantics, although please keep in mind that I have not studied
the implementation of ember-validations so this approach should only be taken if
it makes sense to someone who understands the internals of ember-validations.