Init validations separately from init()
Allows to apply validations to existing model instance.

Useful in situation when you're displaying longer list of items that are very unlikely to be edited - adding validations to everything right away slows it unnecessarily. This way, you can have model without mixin and only when you want to edit it, you apply() the mixin and call initValidations() on it.