centralize and asynchronize the property getter
The base validator is promise-based, but the individual validators are
synchronous. This centralizes the getter into the base validator, waits
for the property to resolve, sync or async, and passes the value into
the individual validators.

clear the errors before populating them

move the assertion check outside of the run

The rest of the tests follow this pattern

remove ic-ajax dependency

It’s not being used and it’s causing installation issues.

add some tests and error handling

update the name of the error handler method

handle validation errors properly

We were switching error contexts. this.errors was being used in most
places and this one used this.model.errors so there were
inconsistencies being created.

add a few more test cases and comments